refactor: separate implementation and test
